refactor(notifications): extract dispatch orchestrator and channel types (PP-734)
